Rename files.py to streams.py
