+++ curl_cffi/requests/session.py
@@ -50,13 +50,19 @@ from .cookies import Cookies, CookieTypes
 from .exceptions import (
     RequestException,
     SessionClosed,
-    UnrewindableBodyError,
     code2error,
 )
-from .files import _AsyncIterableReader
 from .headers import Headers, HeaderTypes
 from .impersonate import BrowserTypeLiteral, ExtraFingerprints, ExtraFpDict
-from .models import STREAM_END, Response
+from .models import Response
+from .streams import (
+    STREAM_END,
+    _AsyncIterableReader,
+    _capture_body_position,
+    _peek_aio_queue,
+    _peek_queue,
+    _rewind_body,
+)
 from .utils import NOT_SET, HttpVersionLiteral, NotSetType, set_curl_options
 from .websockets import (
     AsyncWebSocket,
@@ -189,53 +195,6 @@ def _is_absolute_url(url: str) -> bool:
     """Check if the provided url is an absolute url"""
     parsed_url = urlparse(url)
     return bool(parsed_url.scheme and parsed_url.hostname)
-
-
-def _peek_queue(q: queue.Queue, default=None):
-    try:
-        return q.queue[0]
-    except IndexError:
-        return default
-
-
-def _peek_aio_queue(q: asyncio.Queue, default=None):
-    try:
-        return q._queue[0]  # type: ignore
-    except IndexError:
-        return default
-
-
-_BODY_NOT_REWINDABLE = object()
-
-
-def _capture_body_position(data: RequestData | None, content: object | None):
-    """Capture enough state to safely replay a request body after a failed attempt."""
-    body = content if content is not None else data
-    if not hasattr(body, "read"):
-        if content is not None and isinstance(body, AsyncIterable):
-            return None if body.__aiter__() is not body else _BODY_NOT_REWINDABLE
-        if content is not None and isinstance(body, Iterable):
-            return None if iter(body) is not body else _BODY_NOT_REWINDABLE
-        return None
-    try:
-        if hasattr(body, "seekable") and not body.seekable():
-            raise OSError
-        return body.tell()
-    except (AttributeError, OSError):
-        return _BODY_NOT_REWINDABLE
-
-
-def _rewind_body(body: object | None, position: object) -> None:
-    if position is None:
-        return
-    if position is _BODY_NOT_REWINDABLE:
-        raise UnrewindableBodyError("Unable to rewind streamed request body for retry")
-    try:
-        body.seek(position)
-    except (AttributeError, OSError) as exc:
-        raise UnrewindableBodyError(
-            "Unable to rewind streamed request body for retry"
-        ) from exc
 
 
 RetryBackoff = Literal["linear", "exponential"]
